build: treat warnings as errors, and clear the nine that were already there
- CMAKE_COMPILE_WARNING_AS_ERROR is on by default behind NARRATIVEENGINE_WARNINGS_AS_ERRORS, so another toolchain can still build this.
- The CMake minimum rises to 3.24 because that variable is silently ignored below it, which would look enabled while doing nothing.
- GossipSim's OnLoad and OnSessionEnd locals stop shadowing the globals that alias LIVE state - the shadowing that hid the co-save wipe.
- AliasWalkFilter and FineRoads index BSTArray with its own 32-bit size_type instead of a size_t narrowed at every subscript (C4267).
- LetterComposer's speaker literal becomes speakerRaw, no longer shadowing the raw dialogue response it is parsed out of (C4456).
- CMakePresets rationale moves from //-prefixed cacheVariables into a vendor block, so CMake stops warning about variables it defined.

+++ src/FineRoads.cpp
@@ -186,7 +186,9 @@ namespace NarrativeEngine::FineRoads
 
                 const auto& verts = mesh->vertices;
                 const auto& tris = mesh->triangles;
-                for (std::size_t t = 0; t < tris.size(); ++t) {
+                // 32-bit to match BSTArray's size_type; the triangle index
+                // is narrowed to uint16 for RawNode below either way.
+                for (std::uint32_t t = 0; t < tris.size(); ++t) {
                     const auto& tri = tris[t];
                     if (!tri.triangleFlags.any(TriangleFlag::kPreferred)) {
                         continue;
